changed slider module
