+++ src/js/main.js
@@ -1,7 +1,7 @@
-import {Slider, slider} from "./modules/index";
+import {MainSlider} from "./modules/slider/slider-main";
 
 
 window.addEventListener('DOMContentLoaded', () => {
-    const slider = new Slider('.page', '.next');
+    const slider = new MainSlider({btns: '.next', page: '.page'});
     slider.render();
 });
