+++ src/js/modules/index.js
@@ -1,1 +1,1 @@
-export {Slider} from "./slider";
+export {Slider} from "./slider/slider";
