+++ src/js/modules/slider/slider-main.js
@@ -1,11 +1,11 @@
-export class Slider {
-    constructor(page, btns) {
-        this.page = document.querySelector(page); 
-        this.slides = this.page.children;
-        this.btns = document.querySelectorAll(btns);
-        this.slideIndex = 1;
-    }
+import {Slider} from "./slider";
 
+export class MainSlider extends Slider {
+    constructor(page, btns) {
+        super(page, btns);
+
+    }
+    
     showSlides(n) {
         if (n > this.slides.length) {
             this.slideIndex = 1;
