+++ src/js/modules/slider/slider.js
@@ -1,0 +1,8 @@
+export class Slider {
+    constructor({page = "", btns = "", next = "", prev = ""} = {}) {
+        this.page = document.querySelector(page); 
+        this.slides = this.page.children;
+        this.btns = document.querySelectorAll(btns);
+        this.slideIndex = 1;
+    }
+}
